Introduct PBScalarValueWriter for consistency with PBScalarValueReader
This separates the concerns of writing the tag (PBFieldWriter takes care
of that) with writing the value (now done by PBScalarValueWriter),
because we shouldn't have to know the field index in order to write a
scalar value.

This change brings the reader and writer hierarchies in line with each
other: message -> product -> field -> scalar value.