Skip writing of default scalar values
When writing a scalar field, check if the value is equal to the default
value for its type (e.g. "" for string, 0 for numbers). If so, skip the
field.

We have to be careful *not* to skip list elements and map elements even
if they have the default value.

Also when we implement support for oneof fields, we need to be careful
of those too. From the docs: "If you set a oneof field to the default
value (such as setting an int32 oneof field to 0), the "case" of that
oneof field will be set, and the value will be serialized on the wire."