Always write oneof fields
Don't skip writing a oneof field even if it has a default scalar value.

The implementation for this is a bit nasty. I added a boolean parameter
to the signature of `PBFieldWriter` that is only used when writing a
scalar field and is ignored in all other cases. Will have a think about
a cleaner solution.