Require Coproducts (oneof fields) to be wrapped in Option
This is so we can set the field to None when it is missing when reading
a message from protobuf. If it was a plain coproduct type (e.g.
`Int :+: String :+: CNil`), there is no sensible default value we can
set it to.  We could arbitrarily set it to `Inr(0)`, but we would lose
the information that the field was missing from the message.